7 - add Foler Validations

+++ DogSitterMarketplaceApi/Models/OrdersDto/Response/AvgScoreCommentAboutClientForSitterResponseDto.cs
@@ -4,7 +4,7 @@ namespace DogSitterMarketplaceApi.Models.OrdersDto.Response
 {
     public class AvgScoreCommentAboutClientForSitterResponseDto
     {
-        public double AverageScore { get; set; }
+        public decimal AverageScore { get; set; }
 
         public List<CommentAboutClientsForSitterResponseDto> CommentsAboutClientForSitter { get; set; } = new();
     }
