+++ DogSitterMarketplaceApi/Models/OrdersDto/Response/AvgScoreCommentResponseDto.cs
@@ -4,7 +4,7 @@ namespace DogSitterMarketplaceApi.Models.OrdersDto.Response
 {
     public class AvgScoreCommentResponseDto
     {
-        public double AverageScore { get; set; }
+        public decimal AverageScore { get; set; }
 
         public List<CommentResponseDto> Comments { get; set; } = new();
     }
