+++ DogSitterMarketplaceApi/Models/OrdersDto/Response/AvgScoreCommentWithoutUserResponseDto.cs
@@ -4,7 +4,7 @@ namespace DogSitterMarketplaceApi.Models.OrdersDto.Response
 {
     public class AvgScoreCommentWithoutUserResponseDto
     {
-        public double AverageScore { get; set; }
+        public decimal AverageScore { get; set; }
 
         public List<CommentWithoutUserResponseDto> CommentsWithoutUser { get; set; } = new();
     }
